models/GET/requirements.txt
absl-py==2.0.0
aermanager==0.3.0
ansys-pythonnet==3.1.0rc3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clr-loader==0.2.6
contourpy==1.1.1
cupy @ file:///opt/conda/conda-bld/cupy_1610065936122/work
cycler==0.12.1
Deprecated==1.2.14
distro==1.8.0
dotnetcore2==3.1.23
dv==1.0.12
dv-processing==1.7.9
einops==0.7.0
fastrlock==0.6
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.2
google-auth==2.26.1
google-auth-oauthlib==1.0.0
grpcio==1.60.0
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lxml==5.0.0
lz4==4.3.2
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.7.4
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.23.2
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
plumbum==1.8.2
protobuf==4.25.1
psutil==5.9.7
pyaedt==0.7.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rpyc==5.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shutilwhich==1.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchinfo==1.8.0
torchvision==0.16.0+cu118
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
yacs==0.1.8
zipp==3.17.0
zstd==1.5.5.1
